+++ spec/integration/rails/support/all.rb
@@ -1,0 +1,6 @@
+# frozen_string_literal: true
+
+ENV['RAILS_ENV'] = 'test'
+require_relative 'application'
+require_relative 'home_controller'
+require 'rspec/rails'